issue #80: refactored heating control scheduler to align with hezagonal architecture
