+++ src/main/kotlin/org/agrfesta/sh/api/schedulers/HeatingControlScheduler.kt
@@ -1,18 +1,18 @@
 package org.agrfesta.sh.api.schedulers
 
-import org.agrfesta.sh.api.services.heating.HeatingOrchestrationService
+import org.agrfesta.sh.api.core.application.ports.inbounds.EvaluateHeatingStateUseCase
 import org.springframework.scheduling.annotation.Async
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
 
 @Component
 class HeatingControlScheduler(
-    private val heatingOrchestrationService: HeatingOrchestrationService
+    private val evaluateHeatingState: EvaluateHeatingStateUseCase
 ) {
     @Scheduled(cron = "0 */15 * * * *")
     @Async
     fun scheduledTask() {
-        heatingOrchestrationService.evaluateHeatingState()
+        evaluateHeatingState.execute()
     }
 
 }
