+++ src/main/kotlin/org/agrfesta/sh/api/services/heating/DynamicSharedHeatingStrategyService.kt
@@ -8,6 +8,7 @@ import org.agrfesta.sh.api.core.domain.heating.SharedHeatingStrategy
 import org.agrfesta.sh.api.core.application.ports.outbounds.settings.PropertyRepository
 import org.agrfesta.sh.api.utils.LoggerDelegate
 import org.springframework.beans.factory.annotation.Value
+import org.springframework.context.annotation.Primary
 import org.springframework.stereotype.Service
 
 /**
@@ -18,6 +19,7 @@ import org.springframework.stereotype.Service
  * If the key is missing, the value is invalid, or a persistence error occurs, it falls back to
  * the configured `defaultStrategy` to ensure the system remains operational.
  */
+@Primary
 @Service
 class DynamicSharedHeatingStrategyService(
     @param:Value("\${heating.default-strategy:ECONOMY}") private val defaultStrategy: SharedHeatingStrategy,
